alpha

+++ src/logic/import.rs
@@ -14,7 +14,7 @@ use crate::utils::{
     sql::insert::save_card,
 
 };
-use tui::widgets::ListState;
+use tui::widgets::{ListState, Clear};
 use tui::{
     backend::Backend,
     layout::{Constraint, Direction::{Vertical, Horizontal}, Layout, Rect},
@@ -33,7 +33,7 @@ use crate::MyType;
 use reqwest;
 use std::fs;
 use crate::utils::widgets::ankimporter::{Ankimporter, ShouldQuit};
-use crate::utils::widgets::load_cards::{Template, LoadState};
+use crate::utils::widgets::load_cards::{Template, LoadState, ImportProgress};
 
 
 use sanitize_filename;
@@ -42,11 +42,18 @@ use crate::utils::widgets::filepicker::{FilePicker, PickState};
 use regex::Regex;
 use reqwest::header;
 use std::io::prelude::*;
+use std::sync::mpsc::{Sender, Receiver};
+use std::sync::mpsc;
+use std::thread;
 
 
 fn download_the_deck(url: String) -> String{
-        std::fs::remove_dir_all("./temp").unwrap();
-        std::fs::create_dir("./temp").unwrap();
+        if !std::path::Path::new("./temp").exists(){
+            std::fs::create_dir("./temp").unwrap();
+        } else {
+            std::fs::remove_dir_all("./temp").unwrap();
+            std::fs::create_dir("./temp").unwrap();
+        }
 
         let downloc = "./temp/ankitemp.apkg";
         let body = reqwest::blocking::get(url).unwrap();
@@ -57,14 +64,57 @@ fn download_the_deck(url: String) -> String{
         };
         file.write_all(&(body.bytes().unwrap())).unwrap();
         downloc.to_string()
+}
+
+#[tokio::main]
+pub async fn foo(url: String, transmitter: std::sync::mpsc::SyncSender<(u64, u64)>){
+    download_deck(url, transmitter).await;
+}
+
+use futures_util::StreamExt;
+pub async fn download_deck(url: String, transmitter: std::sync::mpsc::SyncSender<(u64, u64)>){
+    if !std::path::Path::new("./temp").exists(){
+        std::fs::create_dir("./temp").unwrap();
+    } else {
+        std::fs::remove_dir_all("./temp").unwrap();
+        std::fs::create_dir("./temp").unwrap();
     }
+    let downloc = "./temp/ankitemp.apkg";
+
+    let client = reqwest::Client::new();
+    let res = client
+        .get(&url)
+        .send()
+        .await
+        .or(Err(format!("Failed to GET from '{}'", &url))).unwrap();
+
+    let total_size = res
+        .content_length()
+        .ok_or(format!("Failed to get content length from '{}'", &url)).unwrap();
+    
+
+    // download chunks
+    let mut file = File::create(&downloc).or(Err(format!("Failed to create file '{}'",downloc))).unwrap();
+    let mut downloaded: u64 = 0;
+    let mut stream = res.bytes_stream();
+
+    while let Some(item) = stream.next().await {
+        let chunk = item.or(Err(format!("Error while downloading file"))).unwrap();
+        file.write_all(&chunk)
+            .or(Err(format!("Error while writing to file"))).unwrap();
+        let new = std::cmp::min(downloaded + (chunk.len() as u64), total_size);
+        downloaded = new;
+        let _ = transmitter.try_send((new, total_size));
+    }
+
+}
 
 
 
 fn extract_download_link(trd: &String) -> String{
     let pattern = r"(?P<link>(https:.*));".to_string();
     let re = Regex::new(&pattern).unwrap();
-    let foo = re.captures(&trd).unwrap();
+    let foo = re.captures(&trd).expect(&format!("Couldnt find pattern on following string: {}@@", trd));
     foo.get(1).unwrap().as_str().to_string()
 }
 
@@ -75,9 +125,11 @@ fn get_k_value(pagesource: &String) -> String{
     foo.get(1).unwrap().as_str().to_string()
 }
 
-fn get_download_link(pagesource: &String, deck: u32)->String{
-    let k = get_k_value(pagesource);
-    let main = format!("https://ankiweb.net/shared/downloadDeck/{}", deck);
+pub fn get_download_link(deckid: u32)->String{
+    let url = format!("https://ankiweb.net/shared/info/{}", deckid);
+    let pagesource = reqwest::blocking::get(url).unwrap().text().unwrap();
+    let k = get_k_value(&pagesource);
+    let main = format!("https://ankiweb.net/shared/downloadDeck/{}", deckid);
     let body = format!("k={}&submit=Download", k);
     let mut headers = header::HeaderMap::new();
     headers.insert("Content-Type", "application/x-www-form-urlencoded".parse().unwrap());
@@ -92,15 +144,6 @@ fn get_download_link(pagesource: &String, deck: u32)->String{
         .text().unwrap();
     let link = extract_download_link(&result);
     return link;
-}
-
-
-pub fn download_deck(deckid: u32) -> String{
-    let url = format!("https://ankiweb.net/shared/info/{}", deckid);
-    let body = reqwest::blocking::get(url).unwrap().text().unwrap();
-    let download_link = get_download_link(&body, deckid);
-    let download_path = download_the_deck(download_link);
-    download_path
 }
 
 
@@ -131,24 +174,22 @@ impl Selected{
     }
 }
 
-
-
-enum DeckSource{
-    id(u32),
-    path(PathBuf),
+struct Unzipper{
+    rx: mpsc::Receiver<UnzipStatus>,
+    name: String,
 }
 
 
-
+use crate::utils::widgets::load_cards::UnzipStatus;
 enum Menu{
     Main,
     Anki(Ankimporter),
     Local(FilePicker),
     LoadCards(Template),
-    
+    ImportAnki(mpsc::Receiver<ImportProgress>),
+    Unzipping(Unzipper),
 }
 
-//use crate::utils::widgets::filepicker::FilePicker;
 
 pub struct Importer{
     topics: TopicList,
@@ -171,7 +212,7 @@ impl Importer{
             menu,
         }
     }
-    pub fn render(&mut self, f: &mut tui::Frame<MyType>, area: tui::layout::Rect) {
+    pub fn render(&mut self, conn: &Arc<Mutex<Connection>>, f: &mut tui::Frame<MyType>, area: tui::layout::Rect) {
 
         match &mut self.menu{
             Menu::Main => self.render_main(f, area),
@@ -180,11 +221,39 @@ impl Importer{
             },
             Menu::Anki(ankimporter) => {
                 match &ankimporter.should_quit{
-                    ShouldQuit::No => ankimporter.render(f, area),
+                    ShouldQuit::No => ankimporter.render(conn, f, area),
                     ShouldQuit::Yeah => self.render_main(f, area),
-                    ShouldQuit::Takethis(tmp) => self.menu = Menu::LoadCards(tmp.clone()),
+                    ShouldQuit::Takethis(tmp) => {
+                        let foldername = tmp.1.clone();
+                        let dlpath = std::path::Path::new("./temp/ankitemp.apkg").to_path_buf();
+                        let (tx, rx): (mpsc::Sender<UnzipStatus>, Receiver<UnzipStatus>) = mpsc::channel();
+                        thread::spawn( move || {
+                            Template::unzip_deck(dlpath, &foldername, tx);
+                            }
+                        );
+                        self.menu = Menu::Unzipping(
+                            Unzipper{
+                                rx,
+                                name: tmp.1.clone(),
+                            }
+                        );
+                    },
                     };
                 }
+            
+            Menu::Unzipping(unzipper) => {
+                if let Ok(unstat) = unzipper.rx.recv(){
+                    if let UnzipStatus::Ongoing(msg) = unstat{
+                        draw_message(f, area, &msg);
+                    } else{
+                        let tmpl = Template::new(conn, unzipper.name.clone());
+                        self.menu = Menu::LoadCards(tmpl);
+                    }
+                } else {
+                    let tmpl = Template::new(conn, unzipper.name.clone());
+                    self.menu = Menu::LoadCards(tmpl);
+                }
+            }
             
             Menu::LoadCards(tmpl) => {
                 tmpl.render(f, area);
@@ -192,6 +261,36 @@ impl Importer{
                     let aim = Ankimporter::new();
                     self.menu = Menu::Anki(aim);
                 }
+            },
+            Menu::ImportAnki(rx) => {
+                if let Ok(prog) = rx.try_recv(){
+                    let rightcol = Layout::default()
+                        .direction(Vertical)
+                        .constraints(
+                            [
+                            Constraint::Ratio(1, 3),
+                            Constraint::Ratio(1, 3),
+                            Constraint::Ratio(1, 3),
+                            ]
+                            .as_ref(),
+                            )
+                        .split(area);
+                    let mut progbar = rightcol[0];
+
+
+                    progbar.height = std::cmp::min(progbar.height, 5);
+
+                    crate::utils::widgets::progress_bar::progress_bar(f, prog.curr_index as u32, prog.total as u32, Color::LightMagenta, progbar, "Importing cards..");
+                    Field::new_with_text(prog.front, 0, 0).render(f, rightcol[1], false);
+                    Field::new_with_text(prog.back, 0, 0).render(f, rightcol[2], false);
+
+                    if prog.curr_index == prog.total {
+                        self.menu = Menu::Anki(Ankimporter::new());
+                    }
+                } else {
+                        self.menu = Menu::Anki(Ankimporter::new());
+                }
+
             },
         }
     }
@@ -218,6 +317,7 @@ impl Importer{
     pub fn keyhandler(&mut self, conn: &Arc<Mutex<Connection>>, key: MyKey, handle: &rodio::OutputStreamHandle){
         use MyKey::*;
         use Selection::*;
+
         
         match &mut self.menu{
             Menu::Main => self.main_keyhandler(conn, key),
@@ -226,9 +326,9 @@ impl Importer{
                 match &ankimporter.should_quit{
                     ShouldQuit::No => ankimporter.keyhandler(key, conn),
                     ShouldQuit::Yeah => {
-                        self.menu = Menu::Main;
+                        self.menu = Menu::Anki(Ankimporter::new());
                     },
-                    ShouldQuit::Takethis(deck) => {
+                    ShouldQuit::Takethis(_) => {
                     }
                 }
             },
@@ -246,18 +346,35 @@ impl Importer{
                         foldername.pop();
                         foldername.pop();
                         let foldername = foldername.rsplit_once('/').unwrap().1.to_string();
-                        let template = Template::new(conn, path.to_path_buf(), foldername);
+                        let template = Template::new(conn, foldername);
                         self.menu = Menu::LoadCards(template);
                         },
                     };
                 },
             Menu::LoadCards(tmpl) => {
                 match tmpl.state{
-                    LoadState::OnGoing => tmpl.keyhandler(conn, key, handle),
-                    LoadState::Importing(_) => {},
-                    LoadState::Finished => self.menu = Menu::Main,
+                    LoadState::OnGoing => {
+                        tmpl.keyhandler(conn, key, handle);
+                        if let LoadState::Importing = tmpl.state{
+                            let mut tmpclone = tmpl.clone();
+                            let (tx, rx): (mpsc::SyncSender<ImportProgress>, Receiver<ImportProgress>) = mpsc::sync_channel(0);
+                            let connclone = Arc::clone(conn);
+                            thread::spawn(move ||{
+                                tmpclone.import_cards(connclone, tx);
+                            });
+                            self.menu = Menu::ImportAnki(rx);
+
+                        }
+
+                    },
+                    LoadState::Importing => {
+                    },
+                    LoadState::Finished => self.menu = Menu::Anki(Ankimporter::new()),
                 }
             }
+            Menu::ImportAnki(_) => {},
+            Menu::Unzipping(_) => {},
+
         }
     }
 
@@ -271,8 +388,7 @@ impl Importer{
                 self.menu = Menu::Local(fp);
             }, 
             (Anki,  Enter) | (Anki, Char(' ')) => {
-                let aim = Ankimporter::new();
-                self.menu = Menu::Anki(aim);
+                self.menu = Menu::Anki(Ankimporter::new());
             }, 
             (Local, Nav(Direction::Down)) => self.selection = Selection::Anki,
             (Anki,  Nav(Direction::Up))    => self.selection = Selection::Local,
@@ -285,8 +401,4 @@ impl Importer{
 
 }
   
-
-
-
-
